shift to editor service
